parrotMemory is onGetVerbs now in shared
